+++ packages/plugin-video/web/src/tiptap/tiptap.ts
@@ -18,6 +18,9 @@ export const createRicosExtensions: CreateRicosExtensions = defaultOptions => [
     },
     createExtensionConfig: () => ({
       name,
+      group: 'block',
+      selectable: true,
+      draggable: true,
       addOptions: () => defaultOptions,
       addCommands() {
         return {};