+++ packages/tiptap-editor/web/src/models/NodeExtension.tsx
@@ -8,7 +8,12 @@ import {
 } from '@tiptap/react';
 import { pipe } from 'fp-ts/function';
 import React, { ComponentType } from 'react';
-import { DecoratedNodeExtension, DEFAULT_PRIORITY, INodeExtension } from './domain-types';
+import {
+  DecoratedNodeExtension,
+  DEFAULT_PRIORITY,
+  IReactNodeExtension,
+  IHtmlNodeExtension,
+} from './domain-types';
 import { isRicosNodeExtension, RicosExtension, RicosNodeExtension } from 'ricos-tiptap-types';
 import { RicosNode } from '../components/RicosNode';
 
@@ -17,10 +22,6 @@ const toExtensionConfig = (ext: RicosNodeExtension) =>
 
 const toFullNodeConfig = (ext: RicosNodeExtension) => (config: NodeConfig): NodeConfig => ({
   type: 'node',
-  group: 'block',
-  atom: true,
-  selectable: true,
-  draggable: true,
   parseHTML: () => [{ tag: `${config.name}-component` }],
   renderHTML: ({ HTMLAttributes }) => [`${config.name}-component`, mergeAttributes(HTMLAttributes)],
   addAttributes: () => ext.componentDataDefaults || {},
@@ -36,12 +37,12 @@ const createRicosNodeHOC = (Component: ComponentType) => (props: NodeViewRendere
   </NodeViewWrapper>
 );
 
-export class NodeExtension implements INodeExtension {
+export class ReactNodeExtension implements IReactNodeExtension {
   config: NodeConfig;
 
   priority: number;
 
-  type = 'node' as const;
+  type = 'react-node' as const;
 
   name: string;
 
@@ -58,7 +59,7 @@ export class NodeExtension implements INodeExtension {
   }
 
   getComponent() {
-    return this.ricosExtension.Component;
+    return this.ricosExtension.Component as ComponentType;
   }
 
   asRenderable(decoratedComponent: ComponentType) {
@@ -66,13 +67,39 @@ export class NodeExtension implements INodeExtension {
   }
 }
 
-class RenderableNodeExtension extends NodeExtension implements DecoratedNodeExtension {
+class RenderableNodeExtension extends ReactNodeExtension implements DecoratedNodeExtension {
   constructor(component: ComponentType, extension: RicosExtension) {
     super(extension);
     this.config = {
       ...this.config,
       addNodeView: () => pipe(component, createRicosNodeHOC, ReactNodeViewRenderer),
     };
+  }
+
+  toTiptapExtension() {
+    return Node.create(this.config);
+  }
+}
+
+export class HtmlNodeExtension implements IHtmlNodeExtension {
+  config: NodeConfig;
+
+  priority: number;
+
+  type = 'html-node' as const;
+
+  name: string;
+
+  ricosExtension: RicosNodeExtension;
+
+  constructor(extension: RicosExtension) {
+    if (!isRicosNodeExtension(extension)) {
+      throw new TypeError('invalid argument');
+    }
+    this.ricosExtension = extension;
+    this.config = createRicosNodeConfig(extension);
+    this.priority = this.config.priority || DEFAULT_PRIORITY;
+    this.name = this.config.name;
   }
 
   toTiptapExtension() {